+++ vnext/Microsoft.ReactNative.IntegrationTests/ExecuteJsiTests.cpp
@@ -116,7 +116,6 @@ TEST_CLASS (ExecuteJsiTests) {
 
     auto reactNativeHost = TestReactNativeHostHolder(L"ExecuteJsiTests", [](ReactNativeHost const &host) noexcept {
       host.PackageProviders().Append(winrt::make<TestPackageProvider>());
-      return true;
     });
 
     TestEventService::ObserveEvents({