+++ vnext/Microsoft.ReactNative/Modules/I18nManagerModule.h
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 #pragma once
 
+#include "../../codegen/NativeI18nManagerSpec.g.h"
 #include <NativeModules.h>
 #include <winrt/Windows.ApplicationModel.h>
 #include <winrt/Windows.Foundation.h>
@@ -20,8 +21,8 @@ struct I18nManager {
   REACT_INIT(Initialize)
   void Initialize(React::ReactContext const &reactContext) noexcept;
 
-  REACT_CONSTANT_PROVIDER(GetConstants)
-  void GetConstants(React::ReactConstantProvider &provider) noexcept;
+  REACT_GET_CONSTANTS(GetConstants)
+  ReactNativeSpecs::I18nManagerSpec_Constants GetConstants() noexcept;
 
   REACT_METHOD(AllowRTL, L"allowRTL")
   void AllowRTL(bool allowRTL) noexcept;