+++ vnext/codegen/NativeJSDevSupportSpec.g.h
@@ -13,7 +13,18 @@
 
 namespace Microsoft::ReactNativeSpecs {
 
+REACT_STRUCT(JSDevSupportSpec_Constants)
+struct JSDevSupportSpec_Constants {
+    REACT_FIELD(ERROR_CODE_EXCEPTION)
+    double ERROR_CODE_EXCEPTION;
+    REACT_FIELD(ERROR_CODE_VIEW_NOT_FOUND)
+    double ERROR_CODE_VIEW_NOT_FOUND;
+};
+
 struct JSDevSupportSpec : winrt::Microsoft::ReactNative::TurboModuleSpec {
+  static constexpr auto constants = std::tuple{
+      TypedConstant<JSDevSupportSpec_Constants>{0},
+  };
   static constexpr auto methods = std::tuple{
       Method<void(std::string) noexcept>{0, L"onSuccess"},
       Method<void(double, std::string) noexcept>{1, L"onFailure"},
@@ -21,7 +32,14 @@ struct JSDevSupportSpec : winrt::Microsoft::ReactNative::TurboModuleSpec {
 
   template <class TModule>
   static constexpr void ValidateModule() noexcept {
+    constexpr auto constantCheckResults = CheckConstants<TModule, JSDevSupportSpec>();
     constexpr auto methodCheckResults = CheckMethods<TModule, JSDevSupportSpec>();
+
+    REACT_SHOW_CONSTANT_SPEC_ERRORS(
+          0,
+          "JSDevSupportSpec_Constants",
+          "    REACT_GET_CONSTANTS(GetConstants) JSDevSupportSpec_Constants GetConstants() noexcept {/*implementation*/}\n"
+          "    REACT_GET_CONSTANTS(GetConstants) static JSDevSupportSpec_Constants GetConstants() noexcept {/*implementation*/}\n");
 
     REACT_SHOW_METHOD_SPEC_ERRORS(
           0,