+++ packages/react-native-windows-init/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - react-native-windows-init
 
-This log was last generated on Thu, 23 Sep 2021 05:06:48 GMT and should not be manually modified.
+This log was last generated on Wed, 29 Sep 2021 05:06:54 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.63
+
+Wed, 29 Sep 2021 05:06:54 GMT
+
+### Changes
+
+- Bump @react-native-windows/cli to v0.0.0-canary.90 (ngerlem@microsoft.com)
 
 ## 1.1.63